Hide event date on talk pages
Don't render `field_event_date` when viewing a talk node.

It contains the latest event date for that talk, so doesn't need to be
displayed as it's duplication of information that's already displayed
and is only there to make it easier to sort the talks on the Talks
listing page.