Sort talks only by the event date
Sort the talks on the listing page solely by `field_event_date`.

Prior to this commit, the page was using a custom sort plugin that
attempts to show future talks followed by past talks, but also affect
the order to show upcoming talks in an ascending order, followed by past
dalks in the opposite (descending) order.

This was not working as expected.

For now, solely using the event date field is the simpler solution.

In the future, I'll either re-attempt this or remove the code and
simplify my codebase.